@@ -4,16 +4,16 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
   fitness:
     path: ../
-  intl: ^0.16.1
-  lottie: ^0.7.0+1
-  cupertino_icons: ^1.0.0
+  intl: ^0.17.0
+  lottie: ^1.0.1
+  cupertino_icons: ^1.0.3
 
 dev_dependencies:
   flutter_test:
